Solved bugs after refactoring
