+++ src/engine2/models/alembic/env.py
@@ -19,7 +19,7 @@ fileConfig(config.config_file_name)
 # from myapp import mymodel
 # target_metadata = mymodel.Base.metadata
 # ~ target_metadata = None
-from db.desktop import Base
+from models.desktop import Base
 target_metadata = Base.metadata
 
 # other values from the config, defined by the needs of env.py,
