+++ src/engine2/test/grpc/engine_mock.py
@@ -1,10 +1,10 @@
 from common.sm.desktop_sm import DesktopSM
 # ~ import engine.exceptions
-from engine.exceptions import *
+from common.exceptions.engine import *
 
 from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker
-from db.desktop import *
+from models.desktop import *
 from api.grpc.proto import desktop_pb2
 import sys, os
